requirements.txt
python-chess
torch
requests
